+++ src/pages/implementation-references/react-hook-form.tsx
@@ -6,10 +6,14 @@ import { Button, Form } from 'antd';
 import { FileDropzone } from '@/components/forms/file-dropzone';
 import { FileWithPath } from 'file-selector';
 import { PasswordInput } from '@/components/forms/password-input';
+import { SelectInput } from '@/components/forms/select-input';
 interface ISampleForm {
   text: string;
   password: string;
   file: FileWithPath[];
+  simpleSelect: string;
+  multipleSelect: string[];
+  tagsSelect: string[];
 }
 
 const ReactHookFormPage: NextPage = () => {
@@ -35,6 +39,82 @@ const ReactHookFormPage: NextPage = () => {
           label="Senha"
         />
         <FileDropzone<ISampleForm> name="file" label="Arquivo" />
+        <SelectInput<ISampleForm>
+          name="simpleSelect"
+          label="Seletor Simples"
+          options={[
+            {
+              label: 'Option 1',
+              value: 'option 1',
+            },
+            {
+              label: 'Option 2',
+              value: 'option 2',
+            },
+          ]}
+        />
+        <SelectInput<ISampleForm>
+          name="simpleSelect"
+          label="Disabled Select"
+          disabled={true}
+          options={[
+            {
+              label: 'Option 1',
+              value: 'option 1',
+            },
+            {
+              label: 'Option 2',
+              value: 'option 2',
+            },
+          ]}
+        />
+        <SelectInput<ISampleForm>
+          name="simpleSelect"
+          label="Loading Select"
+          loading={true}
+          options={[
+            {
+              label: 'Option 1',
+              value: 'option 1',
+            },
+            {
+              label: 'Option 2',
+              value: 'option 2',
+            },
+          ]}
+        />
+        <SelectInput<ISampleForm>
+          name="multipleSelect"
+          label="Multiple Select"
+          allowClear
+          mode="multiple"
+          options={[
+            {
+              label: 'Option 1',
+              value: 'option 1',
+            },
+            {
+              label: 'Option 2',
+              value: 'option 2',
+            },
+          ]}
+        />
+        <SelectInput<ISampleForm>
+          name="tagsSelect"
+          label="Tags Select"
+          allowClear
+          mode="tags"
+          options={[
+            {
+              label: 'Option 1',
+              value: 'option 1',
+            },
+            {
+              label: 'Option 2',
+              value: 'option 2',
+            },
+          ]}
+        />
         <Button type="primary" htmlType="submit">
           Enviar
         </Button>